// kdiab-measures/frontend/src/api/generated/api.ts
export interface MeasureResponse {
    'id': string;
    'userId': string;
    'measuredAt': string;
    'createdAt': string;
    'type': MeasureType;
    'source': MeasureSource;
    /**
     * Flexible JSON payload whose structure depends on the measure type
     */
    'data': { [key: string]: any; };
    'status': MeasureStatus;
}


/**
 * The origin of the measurement data
 */

export const MeasureSource = {
    Manual: 'MANUAL',
    Nightscout: 'NIGHTSCOUT',
    GoogleFit: 'GOOGLE_FIT',
    AppleHealth: 'APPLE_HEALTH'
} as const;

export type MeasureSource = typeof MeasureSource[keyof typeof MeasureSource];



export const MeasureStatus = {
    Active: 'ACTIVE',
    Archived: 'ARCHIVED'
} as const;

export type MeasureStatus = typeof MeasureStatus[keyof typeof MeasureStatus];
// ...
export const MeasureType = {
    Cgm: 'CGM',
    Bgm: 'BGM',
    BloodPressure: 'BLOOD_PRESSURE',
    Weight: 'WEIGHT',
    Pulse: 'PULSE',
    BgCheck: 'BG_CHECK',
    KetoneCheck: 'KETONE_CHECK'
} as const;

export type MeasureType = typeof MeasureType[keyof typeof MeasureType];